# openapi.yaml
openapi: 3.0.3
info:
  title: 千界花园 (Thousand Realms Garden) API
  description: '群智协同操作系统 REST API — XYZ 三轴自由连接、多 Agent 编排、统一 LLM 适配


    📊 自动扩展: 145 个新端点已从后端路由同步'
  version: 1.0.0
  contact:
    name: Thousand Realms Garden Team
  license:
    name: MIT
servers:
- url: http://localhost:3001/api
  description: 本地开发环境
- url: https://thousand-realms-garden.onrender.com/api
  description: 生产环境 (Render)
...
paths:
...
  /intervention/agent-isolate:
    post:
      tags:
      - Intervention
      summary: POST /intervention/agent-isolate
      operationId: post_intervention_agent-isolate
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /intervention/agent/{id}/emergency-stop:
    post:
      tags:
      - Intervention
      summary: POST /intervention/agent/{id}/emergency-stop
      operationId: post_intervention_agent_id_emergency-stop
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /intervention/agent/{id}/history:
    get:
      tags:
      - Intervention
      summary: GET /intervention/agent/{id}/history
      operationId: get_intervention_agent_id_history
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /intervention/agent/{id}/pause:
    post:
      tags:
      - Intervention
      summary: POST /intervention/agent/{id}/pause
      operationId: post_intervention_agent_id_pause
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /intervention/agent/{id}/resume:
    post:
      tags:
      - Intervention
      summary: POST /intervention/agent/{id}/resume
      operationId: post_intervention_agent_id_resume
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /intervention/agent/{id}/terminate:
    post:
      tags:
      - Intervention
      summary: POST /intervention/agent/{id}/terminate
      operationId: post_intervention_agent_id_terminate
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /intervention/global-pause:
    post:
      tags:
      - Intervention
      summary: POST /intervention/global-pause
      operationId: post_intervention_global-pause
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /intervention/global-resume:
    post:
      tags:
      - Intervention
      summary: POST /intervention/global-resume
      operationId: post_intervention_global-resume
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /intervention/group-reorg:
    post:
      tags:
      - Intervention
      summary: POST /intervention/group-reorg
      operationId: post_intervention_group-reorg
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /intervention/inject-message:
    post:
      tags:
      - Intervention
      summary: POST /intervention/inject-message
      operationId: post_intervention_inject-message
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /intervention/list:
    get:
      tags:
      - Intervention
      summary: GET /intervention/list
      operationId: get_intervention_list
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
  /intervention/request:
    post:
      tags:
      - Intervention
      summary: POST /intervention/request
      operationId: post_intervention_request
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /intervention/rules:
    get:
      tags:
      - Intervention
      summary: GET /intervention/rules
      operationId: get_intervention_rules
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
    post:
      tags:
      - Intervention
      summary: POST /intervention/rules
      operationId: post_intervention_rules
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /intervention/rules/{id}:
    delete:
      tags:
      - Intervention
      summary: DELETE /intervention/rules/{id}
      operationId: delete_intervention_rules_id
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /intervention/tool-review:
    post:
      tags:
      - Intervention
      summary: POST /intervention/tool-review
      operationId: post_intervention_tool-review
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /intervention/{id}/approve:
    post:
      tags:
      - Intervention
      summary: POST /intervention/{id}/approve
      operationId: post_intervention_id_approve
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /intervention/{id}/execute:
    post:
      tags:
      - Intervention
      summary: POST /intervention/{id}/execute
      operationId: post_intervention_id_execute
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /intervention/{id}/reject:
    post:
      tags:
      - Intervention
      summary: POST /intervention/{id}/reject
      operationId: post_intervention_id_reject
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 通过 /auth/login 获取的 JWT Token